allow float values for learning rate cycles

+++ modules/util/config/TrainConfig.py
@@ -712,7 +712,7 @@ class TrainConfig(BaseConfig):
         data.append(("scheduler_params", [], list[dict[str, str]], True))
         data.append(("learning_rate", 3e-6, float, False))
         data.append(("learning_rate_warmup_steps", 200, float, False))
-        data.append(("learning_rate_cycles", 1, int, False))
+        data.append(("learning_rate_cycles", 1, float, False))
         data.append(("epochs", 100, int, False))
         data.append(("batch_size", 1, int, False))
         data.append(("gradient_accumulation_steps", 1, int, False))
